2023 day5

+++ src/main/java/aoc/year2023/day5/Almanac.java
@@ -2,5 +2,5 @@ package aoc.year2023.day5;
 
 import java.util.List;
 
-record Almanac(List<Integer> seeds, List<Garden> gardens) {
+record Almanac(List<Long> seeds, List<Garden> gardens) {
 }
