+++ src/main/java/aoc/year2023/day5/Range.java
@@ -1,4 +1,4 @@
 package aoc.year2023.day5;
 
-record Range(int destinationStart, int sourceStart, int rangeLength) {
+record Range(long destinationStart, long sourceStart, long rangeLength) {
 }
